Get rid of all the warnings
Refactor the LaTeX in order to get rid of hyperref, listings and font warnings.
The side effect is that there are no more Chinese characters in PDF index :).
The only warning left is that of pdflatex complaining about "multiple pdfs with
page group included in a single page", which should be safe to ignore.