Optimize the Banshee algorithm
Make `turn` return just the location-path pairs actually written to the `bests`
array. The former version returned everything that was given to it, making the
list filled with bad paths.

The performance improvement is astonishing -- the castles which were too slow to
process and/or ran out of memory are now calculated almost immediately!

:-)